requirements.txt
arrow==0.15.5
astroid==2.3.3
binaryornot==0.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cookiecutter==1.7.0
future==0.18.2
idna==2.8
isort==4.3.21
Jinja2==2.11.1
jinja2-time==0.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
poyo==0.5.0
pylint==2.4.4
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.7
whichcraft==0.6.1
wrapt==1.11.2
